Get merge building js shell and passing trace/ref tests.  Still need to fix up browser
UltraBlame original commit: bf88f75a66ce7f75b7b558210469ca38296cab96

+++ js/src/jsxml.h
@@ -171,8 +171,9 @@ struct JSXML {
     void                *domnode;       
     JSXML               *parent;
     JSObject            *name;
-    uint32              xml_class;      
-    uint32              xml_flags;      
+    
+    size_t              xml_class;      
+    size_t              xml_flags;      
     union {
         JSXMLListVar    list;
         JSXMLElemVar    elem;
